+++ src/Contracts/Writer/WriterInterface.php
@@ -49,7 +49,7 @@ interface WriterInterface
     public function save(
         ObjectInterface $object,
         PromiseInterface $promise = null,
-        ?bool $prefereRealDateOnUpdate = null,
+        ?bool $preferRealDateOnUpdate = null,
     ): WriterInterface;
 
     /**