+++ src/Writer/MediaWriter.php
@@ -58,9 +58,9 @@ class MediaWriter implements WriterInterface
     public function save(
         ObjectInterface $object,
         PromiseInterface $promise = null,
-        ?bool $prefereRealDateOnUpdate = null,
+        ?bool $preferRealDateOnUpdate = null,
     ): WriterInterface {
-        $this->persist($object, $promise, $prefereRealDateOnUpdate);
+        $this->persist($object, $promise, $preferRealDateOnUpdate);
 
         return $this;
     }