+++ tests/behat/Writer/MyObjectWriter.php
@@ -51,9 +51,9 @@ class MyObjectWriter implements WriterInterface
     public function save(
         ObjectInterface $object,
         PromiseInterface $promise = null,
-        ?bool $prefereRealDateOnUpdate = null,
+        ?bool $preferRealDateOnUpdate = null,
     ): WriterInterface {
-        $this->persist($object, $promise, $prefereRealDateOnUpdate);
+        $this->persist($object, $promise, $preferRealDateOnUpdate);
 
         if ($object instanceof MyObject) {
             $object->saved = 'foo';